+++ cabal-pkg-config-version-hook/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for cabal-pkg-config-version-hook
 
+## Unreleased
+
+* Fix support for Cabal 3.12.1.0, which would otherwise not pass defines to inline C/C++ code (e.g. inline-c-cpp)
+
 ## 0.1.0.1 -- 2023-06-28
 
 * Fix an incomplete pattern match